requirements.txt
aioredis==1.3.1
aiosqlite==0.16.1
arq==0.20
fastapi==0.63.0
paramiko==2.7.2
pydantic==1.8.1
scapy==2.4.5
tortoise-orm==0.17.2
ujson==3.2.0
urllib3==1.26.4
uvicorn==0.13.4
websockets==8.1
